refactor: extract utility functions
